pyproject.toml
[tool.poetry]
name = "abtradebotslavev2"
version = "0.1.0"
description = ""
authors = ["blowyourbra1n <blowyourbra1n@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aiogram = "^3.7.0"
python-binance = "^1.0.19"
loguru = "^0.7.2"
pillow = "^10.3.0"
pyinstaller = "^6.7.0"
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
pybit = "^5.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
